remove rescaling stake

+++ hedera-node/hedera-app/src/main/java/com/hedera/node/app/statedumpers/singleton/BBMStakingInfo.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2024 Hedera Hashgraph, LLC
+ * Copyright (C) 2024-2025 Hedera Hashgraph, LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -28,4 +28,4 @@ public record BBMStakingInfo(
         long unclaimedStakeRewardStart,
         long stake,
         @NonNull long[] rewardSumHistory,
-        int weight) {}
+        long weight) {}
